Cast to int an unsigned int for interoperability with series operators.
git-svn-id: https://piranha-psm.googlecode.com/svn/trunk@257 f32a13d4-301a-0410-ba96-7b22a51b2342

+++ src/math.h
@@ -334,7 +334,7 @@ namespace piranha
         T retval=T(1), a_jm1=T(1), a_j, half_x_pow2=T(natural_pow(2,half_x));
         for (unsigned int j=1;j<=iterations;++j)
         {
-          (a_jm1/=(j*(n+j)))*=-1;
+          (a_jm1/=(int)(j*(n+j)))*=-1;
           a_j=a_jm1;
           a_j*=half_x_pow2;
           retval+=a_j;
